Moved pChart library to common location
pChart is not part of my code and I finally found a way to get it out of
the local code directory.
The library now resides in /common/php/pChart2.1.3
TO get it to work I had to modify the include_path, the GD font path,
and the direct path reference to font names in my code.