Change display logic for history
The date span was in a from/to format which was a pain for stored dates
as it locked in both ends.  The new method shows X number of days before
a given date.  So you can leave the date to float as "today" and change
how many days will be shown by default on load.
Minor update for CSS tab library refresh to latest version