Change daily chart display log
The daily chart was using a 'from date'/'to date' logic.  I've changed
that now to a span and 'to date'.  What this supports is storing the
span in a cookie but leaving the end date to float so that you can
always see the last X number of days (or months or whatever) by default
instead of having to tweak the start date every time.
This change also moves the table style reporting from it's own tab to an
area below the chart style reporting.  The styling still has issues, but
it's a start on simplification of the UI