Second half of rename
A new index.php that redirects to thermo.php.  For the stand alone
thermostat project this is a zero sum change.  But it opens the doors
for the new code project that combines the thermostat and electric
monitoring projects.
And for some reason Git is again confused about the create_schedule.ksh
file (which again has not been changed!)