Set default starting tab
I never could find a way in pure CSS to make one tab selected by
default.  So I had to succumb and use a line of JavaScript to make it
so.
Also, I finally told Git to discard the change to the ksh file that had
never changed so it's stopped barking at me about submitting code.