Beta version of program schedule fetch is live!
My thermostat returns data that is certainly bogus and I'm not sure why,
but it is preventing real useful use of this code.  Perhaps it is my
subscription to e5 ( http://www.earthnetworks.com/e5.aspx ) that is
causing it?
GitHub is incorrectly identifying lines 709 through 719 as changed. They
are not.  I did add line 704 and that might be the cause?