Fix for bug #47
Added a fudge factor to the computation of the number of pixels per
minute.  The bug only showed up when the number of days on the chart was
low.  It disappeared as the number of days increased.  So the fudge
factor scales itself away proportionally.
Added execution time to the logging.