Found and fixed a logic bug
My hosting company moved my accounts to all new servers.  The system
time on the web server was 8 seconds ahead of the time on the MySQL
server.  This exposed a bug in how the timestamps worked in the
half-hourly temperature logging.
Assume webserver system time is 02:00:00.  This makes the MySQL server
time 01:59:52.  So cron on the web server launched the process and the
timestamp is computed form MySQL and the HH:MM:SS is forced to conform
to HH:M0:00 so that the charting code will pick it up.  But that time is
01:50:00 giving the appearance of firing ten minutes early AND
guaranteeing that it will NOT be picked up by the chart.

Because cron does not use the .profile a new config file for shell
scripts is introduced: config.ksh
The cron launch scripts now reference that config file.  With additional
new code to make sure the right directory is used without the need for
hard coding.
The actual modification to the temperature log for this bug is small,
but the previous communication library changes are now taken advantage
of (so this also somewhat deals with issue #16)
Other files have code cleanup and additional comments added.