For issue #46 (pChart security version update)
There is one more step in addition to this code patch.  Go to the
common/php directory and set up a symbolic link to the appropriate
version.
In Unix the command for this is
ln -s pChart2.1.4/ pChart

In Windows the command is
mklink /D pChart pChart2.1.4
You may need to be local machine admin to do this?