No more iframe
The daily temperature table used to use an iframe for display.  It
duplicated the whole page background and was quite ugly.  In the mean
time we figured out how to use Ajax to refresh the chart without a whole
page redraw - so now using the same logic for the table.
draw_daily.php - removed duplicate include of common.php (it's in the
common_chart.php include already)
Removed last vestige of failed experiment to style the input boxes where
OS overrides CSS on appearance of check boxes.
Moved add_include_path to common.php