Large config change here
KLogger.php is a 3rd party library and is now moved to the common tree
(outside of this project).  Documentation about the structure of the
tree is in config.php (should it be in README?)
This results in changes to three configuration files.  Notably
config.php now contains a function that modifies the include path and
performs that addition so that everything that includes config.php picks
up that change