Start of commit for version 3 (this is VERY alpha)
I'm no longer sure if NEST is in or out.  It seemed to yoyo for a while.  I have picked one up though and may start working on that (after version 3 is complete).
If/when I do add NEST, then there will need to be a hardware abstraction layer between the device(s) and the monitor API