Start of commit for version 3 (this is VERY alpha)
Two versions of trying to display electric usage information.  The files with the _dl suffix are the data layer (think V in M) code and the others are the UI portion (think V in MVC).  The C (from MVC) is in the URLs themselves: ugly, but it works.
Requires amCharts to be in your "common" library